+++ src/main/java/roomescape/persistence/JdbcReservationRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.persistence;
 
 import java.time.LocalDate;
 import java.time.LocalTime;
@@ -10,6 +10,11 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Name;
+import roomescape.domain.Reservation;
+import roomescape.domain.ReservationRepository;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.Theme;
 
 @Repository
 public class JdbcReservationRepository implements ReservationRepository {
