+++ src/main/java/roomescape/persistence/JdbcThemeRepository.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.persistence;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -9,6 +9,8 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
+import roomescape.domain.Theme;
+import roomescape.domain.ThemeRepository;
 
 @Repository
 public class JdbcThemeRepository implements ThemeRepository {
