+++ src/main/java/roomescape/service/request/ReservationTimeRequest.java
@@ -1,9 +1,9 @@
-package roomescape.dto;
+package roomescape.service.request;
 
 import jakarta.validation.constraints.NotBlank;
 import java.time.LocalTime;
 import roomescape.domain.ReservationTime;
-import roomescape.handler.TimeValid;
+import roomescape.web.exception.TimeValid;
 
 public record ReservationTimeRequest(@NotBlank(message = "생성할 예약 시간은 필수입니다.") @TimeValid String startAt) {
     public ReservationTime toDomain() {
