+++ src/main/java/roomescape/web/controller/ReservationTimeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.web.controller;
 
 import jakarta.validation.Valid;
 import java.net.URI;
@@ -11,9 +11,9 @@ import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ReservationTimeRequest;
-import roomescape.dto.ReservationTimeResponse;
 import roomescape.service.ReservationTimeService;
+import roomescape.service.request.ReservationTimeRequest;
+import roomescape.service.response.ReservationTimeResponse;
 
 @RestController
 @RequestMapping("/times")
