+++ src/main/java/roomescape/web/controller/ThemeController.java
@@ -1,4 +1,4 @@
-package roomescape.controller;
+package roomescape.web.controller;
 
 import jakarta.validation.Valid;
 import java.net.URI;
@@ -12,9 +12,9 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.dto.ThemeRequest;
-import roomescape.dto.ThemeResponse;
 import roomescape.service.ThemeService;
+import roomescape.service.request.ThemeRequest;
+import roomescape.service.response.ThemeResponse;
 
 @RestController
 @RequestMapping("/themes")
