+++ src/main/java/roomescape/web/exception/response/ErrorResponse.java
@@ -1,11 +1,11 @@
-package roomescape.handler;
+package roomescape.web.exception.response;
 
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
 import org.springframework.validation.FieldError;
 
-record ErrorResponse(String message, List<ErrorDetail> details) {
+public record ErrorResponse(String message, List<ErrorDetail> details) {
     public ErrorResponse(String message) {
         this(message, Collections.emptyList());
     }
