+++ src/test/java/roomescape/integration/ReservationTimeIntegrationTest.java
@@ -12,7 +12,7 @@ import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.server.LocalServerPort;
 import org.springframework.test.context.jdbc.Sql;
-import roomescape.dto.ReservationTimeRequest;
+import roomescape.service.request.ReservationTimeRequest;
 
 /*
  * 테스트 데이터베이스 초기 데이터
