+++ src/test/java/roomescape/persistence/ReservationTimeRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.domain;
+package roomescape.persistence;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
@@ -12,6 +12,8 @@ import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.test.context.jdbc.Sql.ExecutionPhase;
+import roomescape.domain.ReservationTime;
+import roomescape.domain.ReservationTimeRepository;
 
 /*
  * 테스트 데이터베이스 초기 데이터
